[DEV 2.15] Populating content on sessions
The chat POST handler now calls the function
corresponding to the session PATCH endpoints. This
attains population of content onto specific
session objects, based on the `session_id` that
is also visible in the Chat model.

A new endpoint for GETting specific sessions has
been added to retrieve sessions matching the
`session_id` path argument. This will be useful
in file navigation.